+++ common/src/main/kotlin/com/cobblemonrei/StatsInfo.kt
@@ -1,0 +1,9 @@
+package com.cobblemonrei
+
+data class StatsRecipeData(
+    val speciesName: String,
+    val baseStats: Map<String, Int>,
+    val baseStatTotal: Int,
+    val primaryType: String,
+    val secondaryType: String?
+)