+++ common/src/main/kotlin/com/cobblemonrei/jei/pokedex/JeiPokedexInfoRecipe.kt
@@ -1,0 +1,7 @@
+package com.cobblemonrei.jei.pokedex
+
+import com.cobblemonrei.PokedexInfoRecipeData
+
+data class JeiPokedexInfoRecipe(val data: PokedexInfoRecipeData) {
+    val speciesName get() = data.speciesName
+}