+++ common/src/main/kotlin/com/cobblemonrei/jei/stats/JeiStatsRecipe.kt
@@ -1,0 +1,7 @@
+package com.cobblemonrei.jei.stats
+
+import com.cobblemonrei.StatsRecipeData
+
+data class JeiStatsRecipe(val data: StatsRecipeData) {
+    val speciesName get() = data.speciesName
+}